+++ elfo-core/src/utils.rs
@@ -1,3 +1,5 @@
+use std::fmt;
+
 use derive_more::Deref;
 
 #[derive(Debug, Clone, Copy, Default, Hash, PartialEq, Eq, Deref)]
@@ -16,4 +18,19 @@ macro_rules! ward {
     ($o:expr) => (ward!($o, else { return; }));
     ($o:expr, else $body:block) => { if let Some(x) = $o { x } else { $body }; };
     ($o:expr, $early:stmt) => (ward!($o, else { $early }));
+}
+
+pub(crate) struct AlternateForm<T>(pub T);
+
+impl<T: fmt::Display> fmt::Display for AlternateForm<T> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{:#}", self.0)
+    }
+}
+
+#[test]
+fn alternate_form() {
+    let cause = anyhow::anyhow!("inner");
+    let error = cause.context("outer");
+    assert_eq!(format!("{}", AlternateForm(error)), "outer: inner");
 }